replaced int variables with longs

+++ app/src/main/java/com/example/se306_project1/data/ProductsDataProvider.java
@@ -71,8 +71,8 @@ public class ProductsDataProvider {
 
         for (Integer key : brands.keySet()) {
 
-            int productID = key;
-            int categoryID = categories.get(key);
+            long productID = key;
+            long categoryID = categories.get(key);
 
             Double productPrice = prices.get(key);
             String productLongName = longNames.get(key);
